refactor: NavHostViewModel stage 4 - extract favorites state holder
